requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==1.10.21
python-dotenv==1.0.1
sqlalchemy==1.4.49
aiosqlite==0.19.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
email-validator==2.1.0.post1 